+++ mono_repo/CHANGELOG.md
@@ -3,8 +3,10 @@
 - Update to `subosito/flutter-action@v2.4.0`.
 - Require at least Dart >= 2.17.0
 - Add the `list` and `readme` commands.
-- Update github actions to use hashes instead of version numbers
+- Update GitHub actions to use hashes instead of version numbers
 - Add read-only permissions to actions by default
+- Soften detection of a Flutter package. Remove check for `flutter` key in
+  `pubspec.yaml`.
 
 ## 6.2.2